requirements.txt
torch
torchvision
torchaudio
llama-index==0.12.25
llama-index-core==0.12.25
llama-index-node-parser-docling==0.3.1
llama-index-readers-docling==0.3.2
pypdf2>=3.0.1
easyocr>=1.7.2
fastapi
uvicorn
python-multipart
transformers
Pillow
timm 
einops
accelerate>=0.26.0
argparse
regex
numpy
supabase
psycopg2
python-dotenv
vecs
openai
scipy
nltk
pgvector==0.3.2
psycopg2-binary
requests
passlib[bcrypt]
python-jose

